+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Only show real time-warp
 - More stable FPS measurement
 - Add support for rendering of hyperbolic and parabolic orbits
+- Turn rocket engine on and throttle is non-zero
 
 # v0.4.0 - 2019-08-24
 - Clamp camera altitude to [1 mm, 1 Tm]